commenting and other misc small changes
