+++ MogulMoves/app/src/main/java/com/example/mogulmoves/AddBinomialTrialFragment.java
@@ -17,6 +17,8 @@ import androidx.fragment.app.DialogFragment;
 
 import static android.view.View.VISIBLE;
 
+//Fragment to add a new binomial trial.
+
 public class AddBinomialTrialFragment extends DialogFragment {
     private RadioButton success;
     private RadioButton failure;
