+++ MogulMoves/app/src/main/java/com/example/mogulmoves/AddNNCountTrialFragment.java
@@ -16,6 +16,8 @@ import androidx.fragment.app.DialogFragment;
 
 import static android.view.View.VISIBLE;
 
+//Fragment to add a new non-negative count trial.
+
 public class AddNNCountTrialFragment extends DialogFragment {
     private EditText count;
