+++ MogulMoves/app/src/main/java/com/example/mogulmoves/ExperimentList.java
@@ -16,6 +16,8 @@ import static android.view.View.INVISIBLE;
 import static android.view.View.VISIBLE;
 import static com.google.android.gms.common.internal.safeparcel.SafeParcelable.NULL;
 
+//Custom array adapter to display experiment information.
+
 public class ExperimentList extends ArrayAdapter<Experiment> {
 
     private ArrayList<Experiment> experiments;
