+++ MogulMoves/app/src/main/java/com/example/mogulmoves/UserProfilePage.java
@@ -8,6 +8,9 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
+/*Activity to view user profile data. A button allows you to edit it, and another allows you to
+manage QR and bar code data. The latter is not implemented yet.*/
+
 public class UserProfilePage extends AppCompatActivity {
     TextView fullName, email, phone;
     User user;
